tests refactor
