+++ MANIFEST.in
@@ -12,3 +12,4 @@ include *.md
 # Include the data files
 recursive-include data *
 recursive-include src *.py
+recursive-include tests *.py
