+++ src/reoptimizations_algorithms/utils/graph/BaseEdge.py
