+++ src/reoptimizations_algorithms/utils/graph/BaseGraph.py
@@ -1,6 +1,6 @@
 from abc import ABCMeta, abstractmethod
 
-from reoptimizations.utils.graph import Vertex, Edge
+from reoptimizations_algorithms.utils.graph import Vertex, Edge
 
 
 class BaseGraph(metaclass=ABCMeta):
