+++ src/reoptimizations_algorithms/utils/graph/UndirectedGraph.py
@@ -1,8 +1,8 @@
 from typing import Dict, List
 
-from reoptimizations.utils.graph import Edge
-from reoptimizations.utils.graph.Graph import Graph
-from reoptimizations.utils.graph.Vertex import Vertex
+from reoptimizations_algorithms.utils.graph import Edge
+from reoptimizations_algorithms.utils.graph.Graph import Graph
+from reoptimizations_algorithms.utils.graph.Vertex import Vertex
 
 
 class UndirectedGraph(Graph):
