+++ src/reoptimizations_algorithms/utils/graph/Vertex.py
@@ -1,6 +1,6 @@
 from typing import Dict
 
-from reoptimizations.utils.graph.Edge import Edge
+from reoptimizations_algorithms.utils.graph.Edge import Edge
 
 
 class Vertex:
