+++ tox.ini
@@ -20,11 +20,13 @@ deps =
     twine
     pytest
 commands =
-    check-manifest --ignore tox.ini,src/tests*
+    check-manifest --ignore tox.ini,tests*
     python setup.py sdist
     twine check dist/*
     flake8 src/
     py.test {posargs}
+setenv =
+    PYTHONPATH = src
 
 [flake8]
 exclude = .tox,*.egg,build,data
